Fix dask 2024.11+ incompatibility
Using ODC with Dask 2024.11+ made it impossible to load data, with
failures looking like `.nodata` attribute found, or other similarly
unhelpful messages.

This was happening because the `datacube.model.Measurement` class, which
contains important values like .dtype and .nodata, wasn't being
correctly serialised/deserialised in those dask versions. Dask switched
to using custom types for building the Dask Graph, but in the process
added a special case for `dict`, which is fine, and for
`dict-subclasses` which is more questionable. This converted all the ODC
Measurement classes into `dict`s, which couldn't be used in the same
way.

I think it should be changed in dask, but haven't started that
conversationn yet. For now, I've changed the Measurement class to no
longer subclass `dict`, and instead implemented all the dunder methods
to make it behave as a dict, while not actually being one.

Replace Measurement class