Hardcode to ns precision times in xarray
xarray 2025+ added support for timecoordinates not being stored to
nanosecond precision. This is lovely, and we probably should support it,
but it broke our tests. I'm hard coding back to nanoseconds for now, in
case there's other downstream implications or assumptions that might
break.

I'm not sure, but the only reason I can see to support