Tests pass as root, fix possible crash in config loading
The config file tests were failing if run as root. They assumed
that /etc/shadow would not be readable. Although running as root is
generally a terrible idea, doing so within a container is relatively
common and sometimes hard to work around.

The config file loader would crash if provided with the name of a
directory as one of the config files to attempt to load.