Improve automated release process with OIDC and uv
Build packages directly on the GHA runner, instead of inside the docker
container in the GHA runner.

Use uv to build the packages, instead of directly running setup.py which
is no longer recommneded.

Use the PyPA GHA Action to upload the package, authenticating with OIDC.

No longer requires a stored authentication key for publishing.